api/requirements.txt
fastapi==0.115.0
pydantic>=2.10.0,<3
uvicorn[standard]==0.30.6
openai>=1.0.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-api-python-client>=2.144.0
supabase>=2.0.0

